# backend/generated/src/main/java/com/recyclegang/backend/generated/api/ReservationsApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.15.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.recyclegang.backend.generated.api;

import com.recyclegang.backend.generated.model.Error;
import com.recyclegang.backend.generated.model.Photo;
import com.recyclegang.backend.generated.model.Reservation;
import com.recyclegang.backend.generated.model.ReservationInput;
import com.recyclegang.backend.generated.model.Ticket;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.lang.Nullable;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.15.0")
@Validated
public interface ReservationsApi {

    public static final String PATH_CANCEL_RESERVATION = "/api/consumer/v1/reservations/{reservationId}/cancel";
    /**
     * POST /api/consumer/v1/reservations/{reservationId}/cancel
     *
     * @param reservationId  (required)
     * @return Success (status code 200)
     *         or Business or validation error (status code 200)
     */
    @RequestMapping(
        method = RequestMethod.POST,
        value = ReservationsApi.PATH_CANCEL_RESERVATION,
        produces = { "application/json" }
    )

    ResponseEntity<Reservation> cancelReservation(
         @PathVariable("reservationId") String reservationId
    );


    public static final String PATH_COMPLETE_DROPOFF = "/api/consumer/v1/reservations/{reservationId}/complete";
    /**
     * POST /api/consumer/v1/reservations/{reservationId}/complete
     *
     * @param reservationId  (required)
     * @return Success (status code 200)
     *         or Business or validation error (status code 200)
     */
    @RequestMapping(
        method = RequestMethod.POST,
        value = ReservationsApi.PATH_COMPLETE_DROPOFF,
        produces = { "application/json" }
    )

    ResponseEntity<Reservation> completeDropoff(
         @PathVariable("reservationId") String reservationId
    );


    public static final String PATH_CREATE_RESERVATION = "/api/consumer/v1/reservations";
    /**
     * POST /api/consumer/v1/reservations
     *
     * @param reservationInput  (required)
     * @return Success (status code 201)
     *         or Business or validation error (status code 200)
     */
    @RequestMapping(
        method = RequestMethod.POST,
        value = ReservationsApi.PATH_CREATE_RESERVATION,
        produces = { "application/json" },
        consumes = { "application/json" }
    )

    ResponseEntity<Reservation> createReservation(
         @Valid @RequestBody ReservationInput reservationInput
    );


    public static final String PATH_DOWNLOAD_PHOTO = "/api/consumer/v1/reservations/{reservationId}/photos/{photoId}";
    /**
     * GET /api/consumer/v1/reservations/{reservationId}/photos/{photoId}
     *
     * @param reservationId  (required)
     * @param photoId  (required)
     * @return JPEG or PNG (status code 200)
     *         or Error (status code 200)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = ReservationsApi.PATH_DOWNLOAD_PHOTO,
        produces = { "application/octet-stream", "application/json" }
    )

    ResponseEntity<org.springframework.core.io.Resource> downloadPhoto(
         @PathVariable("reservationId") String reservationId,
         @PathVariable("photoId") String photoId
    );


    public static final String PATH_GET_RESERVATION = "/api/consumer/v1/reservations/{reservationId}";
    /**
     * GET /api/consumer/v1/reservations/{reservationId}
     *
     * @param reservationId  (required)
     * @return Success (status code 200)
     *         or Business or validation error (status code 200)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = ReservationsApi.PATH_GET_RESERVATION,
        produces = { "application/json" }
    )

    ResponseEntity<Reservation> getReservation(
         @PathVariable("reservationId") String reservationId
    );


    public static final String PATH_GET_TICKET = "/api/consumer/v1/reservations/{reservationId}/ticket";
    /**
     * GET /api/consumer/v1/reservations/{reservationId}/ticket
     *
     * @param reservationId  (required)
     * @return Success (status code 200)
     *         or Business or validation error (status code 200)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = ReservationsApi.PATH_GET_TICKET,
        produces = { "application/json" }
    )

    ResponseEntity<Ticket> getTicket(
         @PathVariable("reservationId") String reservationId
    );


    public static final String PATH_LIST_RESERVATIONS = "/api/consumer/v1/reservations";
    /**
     * GET /api/consumer/v1/reservations
     *
     * @return Success (status code 200)
     *         or Business or validation error (status code 200)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = ReservationsApi.PATH_LIST_RESERVATIONS,
        produces = { "application/json" }
    )

    ResponseEntity<List<Reservation>> listReservations(

    );


    public static final String PATH_UPLOAD_PHOTO = "/api/consumer/v1/reservations/{reservationId}/photos";
    /**
     * POST /api/consumer/v1/reservations/{reservationId}/photos
     *
     * @param reservationId  (required)
     * @param file  (required)
     * @return Success (status code 201)
     *         or Business or validation error (status code 200)
     */
    @RequestMapping(
        method = RequestMethod.POST,
        value = ReservationsApi.PATH_UPLOAD_PHOTO,
        produces = { "application/json" },
        consumes = { "multipart/form-data" }
    )

    ResponseEntity<Photo> uploadPhoto(
         @PathVariable("reservationId") String reservationId,
         @RequestPart(value = "file", required = true) MultipartFile file
    );

}