# contracts/openapi/consumer.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "Recycle Gang consumer API",
    "version": "0.1.0",
    "description": "Consumer and facility admission contract. Local demo credentials are restricted to the local Spring profile. Prices and payments are fixtures."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/consumer/v1/catalog": {
      "get": {
        "operationId": "getCatalog",
        "tags": [
          "Catalog"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Catalog"
                }
              }
            }
          },
          "default": {
            "description": "Business or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/consumer/v1/profile": {
      "get": {
        "operationId": "getProfile",
        "tags": [
          "Profile"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          },
          "default": {
            "description": "Business or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "updateProfile",
        "tags": [
          "Profile"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          },
          "default": {
            "description": "Business or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Profile"
              }
            }
          }
        }
      }
    },
    "/api/consumer/v1/reservations": {
      "get": {
        "operationId": "listReservations",
        "tags": [
          "Reservations"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Reservation"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Business or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createReservation",
        "tags": [
          "Reservations"
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reservation"
                }
              }
            }
          },
          "default": {
            "description": "Business or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservationInput"
              }
            }
          }
        }
      }
    },
    "/api/consumer/v1/reservations/{reservationId}": {
      "get": {
        "operationId": "getReservation",
        "tags": [
          "Reservations"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reservation"
                }
              }
            }
          },
          "default": {
            "description": "Business or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "reservationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/consumer/v1/reservations/{reservationId}/cancel": {
      "post": {
        "operationId": "cancelReservation",
        "tags": [
          "Reservations"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reservation"
                }
              }
            }
          },
          "default": {
            "description": "Business or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "reservationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/consumer/v1/reservations/{reservationId}/ticket": {
      "get": {
        "operationId": "getTicket",
        "tags": [
          "Reservations"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ticket"
                }
              }
            }
          },
          "default": {
            "description": "Business or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "reservationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/consumer/v1/reservations/{reservationId}/complete": {
      "post": {
        "operationId": "completeDropoff",
        "tags": [
          "Reservations"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reservation"
                }
              }
            }
          },
          "default": {
            "description": "Business or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "reservationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/consumer/v1/reservations/{reservationId}/photos": {
      "post": {
        "operationId": "uploadPhoto",
        "tags": [
          "Reservations"
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Photo"
                }
              }
            }
          },
          "default": {
            "description": "Business or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "reservationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        }
      }
    },
    "/api/consumer/v1/reservations/{reservationId}/photos/{photoId}": {
      "get": {
        "operationId": "downloadPhoto",
        "tags": [
          "Reservations"
        ],
        "parameters": [
          {
            "name": "reservationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "photoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JPEG or PNG",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "Error": {
        "$ref": "./shared.yaml#/components/schemas/Error"
      },
      "Facility": {
        "$ref": "./shared.yaml#/components/schemas/Facility"
      },
      "ItemType": {
        "$ref": "./shared.yaml#/components/schemas/ItemType"
      },
      "Slot": {
        "$ref": "./shared.yaml#/components/schemas/Slot"
      },
      "Catalog": {
        "$ref": "./shared.yaml#/components/schemas/Catalog"
      },
      "Address": {
        "$ref": "./shared.yaml#/components/schemas/Address"
      },
      "Profile": {
        "$ref": "./shared.yaml#/components/schemas/Profile"
      },
      "ItemInput": {
        "$ref": "./shared.yaml#/components/schemas/ItemInput"
      },
      "ReservationInput": {
        "$ref": "./shared.yaml#/components/schemas/ReservationInput"
      },
      "ReservationItem": {
        "$ref": "./shared.yaml#/components/schemas/ReservationItem"
      },
      "Photo": {
        "$ref": "./shared.yaml#/components/schemas/Photo"
      },
      "Reservation": {
        "$ref": "./shared.yaml#/components/schemas/Reservation"
      },
      "Ticket": {
        "$ref": "./shared.yaml#/components/schemas/Ticket"
      },
      "AdmissionInput": {
        "$ref": "./shared.yaml#/components/schemas/AdmissionInput"
      },
      "Admission": {
        "$ref": "./shared.yaml#/components/schemas/Admission"
      }
    }
  }
}